requirements.txt
streamlit>=1.36,<2.0

pandas>=2.1,<3.0
numpy>=1.26,<3.0

scikit-learn>=1.4,<2.0
joblib>=1.3,<2.0
matplotlib>=3.8,<4.0

sentence-transformers>=3.0,<4.0
transformers>=4.41,<5.0
huggingface-hub>=0.34,<1.0

pypdf>=4.2,<5.0
faiss-cpu>=1.8,<2.0
xgboost>=2.0,<3.0
